Se agregaron las categorias para poder practicar las relaciones entre entidades y como conectarlas con Java

+++ src/main/java/com/example/demo/repository/ProductRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repository;
 import com.example.demo.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {}
+public interface ProductRepository extends JpaRepository<Product, Long> {}
